Improve performance of role dependency serialization logic
Ansible serializes roles to send them to worker processes. Roles and their
dependencies form a tree of nodes that gets serialized and then sent to the
worker process as a whole.

If one role is present several times in the dependency cycle it gets serialized
several times. This can cause huge memory consumption if you make heavy use of
role dependencies.

This is a change to the serialization logic to reduce the memory consumption. It
changes the serialization to serialize all nodes once, then serialize a tree of
references. If a role is present several times, only the references are
duplicated, not the entire serialized role.