Speed up triggering handlers and listeners via `notify`
After a `notify`, to find the handlers to invoke Ansible iterates through all
defined handlers, runs their names through the template engine. It does so
because handler names can contain variables. Afterwards it checks if the name
matches the handler to be notified.

Running the names through the template engine is a very expensive operation. It
first reads all variables from the inventory, including `group_vars` and
`host_vars`, merges them with the variables set in the current play, assembles
the resulting variables and their values in one big `dict`, and passes this
`dict` to the templating engine for variable substitution.

This does not scale well at all: 1) the more content there is in `group_vars`
and `host_vars`, the longer it takes to read the variables from the inventory.
2) the more handlers are defined, the longer it takes to find matching handlers.
The worst part is, the runtime scales by the multiple of 1) and 2) due to the
loop-within-a-loop structure of the code.

To address that, this patch adds short-circuit logic that skips the
initialization and use of the templating engine if the name of the handler does
not seem to contain variables (it looks for '{{').

This does not fix the playbook runtime scaling issues of `notify` completely,
but it does significantly speed up `notify` as long as there are very few
handlers defined that contain variables in their name.