I FOUND THE BUG
The reason that the "obvious" fix of taking away some of the elements of
the hash doesn't work is because deep in the Ansible the
`_allow_duplicates` flag has the wrong default and this logic only gets
triggered sometimes. I was able to get it to trigger by mixing playbook
roles and includes

roles:
  - example
  - exmaple
tasks:
  - include_role: name=example

runs three times because of this bug.