Show correct changed output in check mode
Right now the command module has silly behavior in check mode. If the
command is skipped because of a when block or something it shows as
skipped. But if the command would actually run it ALSO shows as skipped
despite the fact that the module knows that the command would run.

I'm not the only one annoyed by this either:
Fixes: https://github.com/ansible/ansible/issues/14075
       https://github.com/ansible/ansible/issues/9508

Don't tell me that command doesn't support check mode when
supports_check_mode=True.